+++ vscode-extension/src/noraops/runner/catalogClient.js
@@ -1,9 +1,9 @@
 const { getNoraOpsConfig } = require("../config");
 const { fetchPublishedCatalog } = require("../portalApi");
 
-async function searchPublishedApps(query) {
+async function searchPublishedApps(query, options = {}) {
   const { serverBaseUrl } = getNoraOpsConfig();
-  return fetchPublishedCatalog(serverBaseUrl, query || "");
+  return fetchPublishedCatalog(serverBaseUrl, query || "", options);
 }
 
 module.exports = { searchPublishedApps };